repairmate_backend/requirements.txt
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.6.0
python-decouple==3.8
requests==2.31.0 
pytz==2024.2

#pip install django-storages boto3
#pip install python-magic